# Conflicts:
#	pom.xml

+++ src/main/java/ru/beeline/techradar/repository/TechVersionRepository.java
@@ -5,7 +5,7 @@ import ru.beeline.techradar.domain.TechVersion;
 
 import java.util.List;
 
-public interface TechVersionRepository extends JpaRepository<TechVersion, Long> {
+public interface TechVersionRepository extends JpaRepository<TechVersion, Integer> {
     TechVersion findByTechIdAndIdAndDeletedDateIsNull(Integer techId, Integer id);
     List<TechVersion> findAllByTechIdAndDeletedDateIsNull(Integer techId);
 }
